refactor: split MVC classes of Activity Log page
